fix bugs & refactor

+++ scrapper/src/main/resources/db/changelog/changeset/init/create-chat-table.sql
@@ -1,6 +1,6 @@
 --liquibase formatted sql
 --changeset mmstr:22.02.2024-create-chat-table
 create table if not exists chat(
-    id int primary key,
-    chat_id int unique
+    id bigint primary key,
+    chat_id bigint unique
 );
